+++ include/math/ds/KNN.h
@@ -21,6 +21,8 @@
  */
 
 #include "../numerics/utils.h"
+#include "../Matrix.h"
+#include "../matrix_utils.h"
 #include "Classifier.h"
 #include <cstddef>
