+++ include/math/ds/utils.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "../Matrix.h"
+#include "../matrix_utils.h"
 
 double relu(const double& in) { return std::max(0.0, in); }
 double sigmoid(const double& in) { return 1.0 / (1.0 + exp(-in)); }
