+++ include/math/numerics/lin_alg/newton.h
@@ -26,7 +26,7 @@ using LinearEquation = std::function<Matrix<double>(const Matrix<double>&)>;
  * @param x0 start value
  * @param TOL desired tolerance of the method
  * @param maxIter maximum iterations for the method
- * @return approximated values paired with required number iterations for given tolerance
+ * @returns approximated values paired with required number iterations for given tolerance
  */
 std::pair<Matrix<double>, int>
 newton(const LinearEquation& f, const Jacobian& Df, const Matrix<double>& x0, double TOL, int maxIter) {
