+++ include/math/numerics/ode/odeBDF2.h
@@ -24,7 +24,7 @@
  * @param tInterval interval to perform approximation on
  * @param y0 start value
  * @param option solver options
- * @return approximated values
+ * @returns approximated values
  */
 ODEResult
 ODEBDF2(const ODE& fun, const std::vector<double>& tInterval, const Matrix<double>& y0, const ODEOption& option) {
