+++ include/math/statistics/Insurance.h
@@ -27,7 +27,7 @@ struct Interest {
  * Compounding factor
  * $$1+i$$
  * @param i interest rate
- * @return 1+i
+ * @returns 1+i
  */
 double compoundingFactor(double i);
 /**
@@ -35,54 +35,54 @@ double compoundingFactor(double i);
  * $$\frac{1}{1+i}$$
  *
  * @param i interest rate
- * @return 1/(1+i)
+ * @returns 1/(1+i)
  */
 double discountFactor(double i);
 
 /**
  * calculates the fundamental value, the starting balance
  * @param I values to compute from
- * @return fundamental balance
+ * @returns fundamental balance
  */
 double fundamentalValue(Interest I);
 /**
  * calculates the end value, the balance at the end of I.n periods
  * @param I values to compute from
- * @return
+ * @returns
  */
 double endValue(Interest I);
 /**
  * calculates the term per period of given interest
  * @param I values to compute from
- * @return
+ * @returns
  */
 double termInPeriods(Interest I);
 /**
  * calculates the fundamental value, the starting balance
  * for annuity in advance.
  * @param I values to compute from
- * @return
+ * @returns
  */
 double fundamentalValueOfAnnuityInAdvance(Interest I);
 /**
  * calculates the end value, the balance at the end of I.n periods
  * for annuity in advance
  * @param I values to compute from
- * @return
+ * @returns
  */
 double endValueOfAnnuityInAdvance(Interest I);
 /**
  * calculates the fundamental value, the starting balance
  * for annuity in arrear.
  * @param I values to compute from
- * @return
+ * @returns
  */
 double fundamentalValueOfAnnuityInArrear(Interest I);
 /**
  * calculates the end value, the balance at the end of I.n periods
  * for annuity in arrear.
  * @param I values to compute from
- * @return
+ * @returns
  */
 double endValueOfAnnuityInArrear(Interest I);
